forms.py: for prescription creation, now takes in the patient uuid so that the doctor does not need to remember what patient the prescription is created for
models.py:

urls.py: needed to change the url for the prescription creation to reflect that the patient uuid was passed in

views.py: rx creation needed to be changed to reflect the fact that patient uuid was being passed in. Also, added error that will be displayed on the page if the doctor tries to make a prescription with a start date that is in the past

appt_create.html: now displays the error in a bold, red color

rx_create.html: now displays the error in a bold, red color

rx_delete.html: now does not redirect to appt calendar after clicking cancel. redirects to the user's home page

patient_viewing.html: removed the rx_prescription in page form, added a delete column to the prescription list