forms.py: created a patient discharge form and fixed the patient admit form to the correct fields to be included and excluded
models.py: changes to the patient model where I added new defs to the patient to determine whether or not a patient is requested to be transfer and if a patient is admitted

patient_admit.html: added place to display the patient name

patient_discharge.html: a page to display the discharge form. it's very similar to the prescription deletion form

patient_viewing.html: now displays the admit status underneath the profile picture and will display an admit button if the patient is not admitted. If the patient is admitted, it will display the current hospital

user_doctor.html: the discharge tab now only shows the patients with that doctor as a provider and if the patient is actually admitted. the transfer request only shows the patients with the doctor as a provider and if the patient has a transfer request

urls.py: added url for patient discharge and changed the patient admit url to include patient uuid

view.py: added a patient discharge view and edited the patient admit view to change the patient admit status