Medical Conditions Creation related changes
admin.py: added Medical Conditions to be viewable on admin console
forms.py: added Medical Condition Add form
urls.py: added the url for Medical Condition Add
views.py: added the view for Medical Condition Add

mc_add.html: added the template for adding a medical condition
user_patient.html: added a button to view the button for adding a medical condition. This is not where it should be because a patient can click on it but there's not current staff view of patient, so I need to be able to get to it to test it. Currently, a patient clicking on it will see that they do not have perimission.