Admit changes
rules.py:
more rules for checking if a doctor or nurse has perm to attend a patient
more rules for checking if a patient is admitted

base_user.html:
if the profile being visited is a patient and if the visiter is a doctor or nurse
of that patient, check to see if the patient is not admitted. if the patient is
not admitted, display a button for admitting
if the profile being visited is a patient and if the visiter is a doctor or nurse
of that patient, check to see if the patient is admitted. if the patient is
admitted, display that they are admitted and the location

user_nurse.html:
using user instead of patient_viewing

user_doctor.html:
using user instead of patient_viewing

models.py:
appointment has new def to see if it is occuring today

prescription_list.html:
displays medical conditions and medical history items too

urls.py:
readded the rx_create url (I know that there is the rx_op url but I don't
know how to use it)