Changes to exporting information. Now, with privacy disclaimer and a validiation form checking sec. question. Changes to the base_user so a nurse or doctor can click an admit button from the patient profile.
security_question.html: the template for the security_question form

user_patient.html: better formatting on the indentations

base_user.html: admit button on patient profiles when a doctor or nurse are viewing

views.py: a view for the security question form

urls.py: changing the export information to go to security question first

rules.py: more rules for the admit button

forms.py: the ser. question now has a form. It's not a modelsForm.